Add files via upload
Recommendation system help users to find and select items(e.g. books, movies, restaurants) from the huge number available on the web or in other electronic sources. The purpose of a recommendation system basically is to search for content that would be interesting to an individual. Recommendation systems are Artificially Intelligence based algorithms that skim through all possible options and create a customized list of items that are interesting and relevant to an individual. This results are based on their profile, search/browsing history, what other people with similar traits are watching. So I have build a content based and some kind of popularity based movie  recommendation system using cosine similarity algorithm.